chore: refactor utility classes into utility package for organization
